+++ src/main/java/de/maxhenkel/audioplayer/AudioPlayerPermissionManager.java
@@ -7,7 +7,6 @@ import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.server.level.ServerPlayer;
 
-import javax.annotation.Nullable;
 import java.util.List;
 
 public class AudioPlayerPermissionManager implements PermissionManager<CommandSourceStack> {
@@ -34,14 +33,13 @@ public class AudioPlayerPermissionManager implements PermissionManager<CommandSo
             if (!p.permission.equals(permission)) {
                 continue;
             }
+            if (!p.canUse()) {
+                return false;
+            }
             if (stack.isPlayer()) {
                 return p.hasPermission(stack.getPlayer());
             }
-            if (p.getType().equals(PermissionType.OPS)) {
-                return stack.hasPermission(2);
-            } else {
-                return p.hasPermission(null);
-            }
+            return stack.hasPermission(2);
         }
         return false;
     }
@@ -67,27 +65,24 @@ public class AudioPlayerPermissionManager implements PermissionManager<CommandSo
             this.type = type;
         }
 
-        public boolean hasPermission(@Nullable ServerPlayer player) {
+        public boolean canUse() {
+            return true;
+        }
+
+        public boolean hasPermission(ServerPlayer player) {
             if (isFabricPermissionsAPILoaded()) {
                 return checkFabricPermission(player);
             }
             return type.hasPermission(player);
         }
 
-        private boolean checkFabricPermission(@Nullable ServerPlayer player) {
-            if (player == null) {
-                return false;
-            }
+        private boolean checkFabricPermission(ServerPlayer player) {
             TriState permissionValue = Permissions.getPermissionValue(player, permission);
-            switch (permissionValue) {
-                case DEFAULT:
-                    return type.hasPermission(player);
-                case TRUE:
-                    return true;
-                case FALSE:
-                default:
-                    return false;
-            }
+            return switch (permissionValue) {
+                case DEFAULT -> type.hasPermission(player);
+                case TRUE -> true;
+                default -> false;
+            };
         }
 
         public PermissionType getType() {
@@ -102,8 +97,8 @@ public class AudioPlayerPermissionManager implements PermissionManager<CommandSo
         }
 
         @Override
-        public boolean hasPermission(@Nullable ServerPlayer player) {
-            return AudioPlayer.SERVER_CONFIG.allowStaticAudio.get() && super.hasPermission(player);
+        public boolean canUse() {
+            return AudioPlayer.SERVER_CONFIG.allowStaticAudio.get() && super.canUse();
         }
     }
 
@@ -111,7 +106,7 @@ public class AudioPlayerPermissionManager implements PermissionManager<CommandSo
 
         EVERYONE, NOONE, OPS;
 
-        boolean hasPermission(@Nullable ServerPlayer player) {
+        boolean hasPermission(ServerPlayer player) {
             return switch (this) {
                 case EVERYONE -> true;
                 case NOONE -> false;